Update weight calculations and schema for shipping modules
Refactored weight and chargeable weight calculations in ShippingInstructionController and SalesorderController to use correct bracket selection logic and include pieces in calculations. Improved error handling for missing routes and weight brackets. Updated DateHelper to use Carbon for date parsing. Modified CostController to include type_cost_initial in cost queries. Removed 'type' validation from VendorController's updateVendor method. Changed gross_weight and chargeable_weight columns to decimal in AWB migration for better precision.

+++ database/migrations/2025_07_14_082432_awb.php
@@ -25,8 +25,8 @@ return new class extends Migration
             $table->integer('pol')->unsigned();
             $table->integer('pod')->unsigned();
             $table->string('commodity')->nullable();
-            $table->integer('gross_weight')->unsigned();
-            $table->integer('chargeable_weight')->unsigned();
+            $table->decimal('gross_weight', 10, 2)->unsigned();
+            $table->decimal('chargeable_weight', 10, 2)->unsigned();
             $table->integer('pieces')->unsigned();
             $table->text('special_instructions')->nullable();
             $table->integer('created_by')->unsigned();
